progress

+++ src/CoreEngines/EntityTrackerScreenGameState.cpp
@@ -113,7 +113,7 @@ void nb::EntityTrackerScreenGameState::track( const Entity* entity )
 		auto val = health->getHealth();
 		m_playerHealthValueLabel->setText( to_string( val ) );
 		health->s_onChange.connect_track( m_connections, [&] ( const HealthComponent* const comp,
-															   int change ){
+															   float change ){
 			m_playerHealthValueLabel->setText( to_string( comp->getHealth() ) + " / " + to_string( comp->getMaxHealth() ) );
 		} );
 	}
